requirements.txt
kivy>=2.2.0
Pillow>=10.0.0
openai>=1.0.0
plyer>=2.1.0
